refactored database fixtures
