+++ tests/fixtures/__init__.py
@@ -1,2 +1,3 @@
-from .clients.mealie.fixture_mealie_client import *
+from .databases.aws.fixture_core_database import *
+from .databases.mealie.fixture_mealie_database import *
 from .fixture_users import *
